增加jwt和错误处理

+++ src/controllers/user.ts
@@ -1,6 +1,7 @@
 // src/controllers/user.ts
 import { getManager } from "typeorm";
 import { User } from "../entity/user";
+import { NotFoundException, ForbiddenException } from "../exceptions";
 
 export default class UserController {
   public static async listUsers(ctx: any) {
@@ -19,11 +20,17 @@ export default class UserController {
       ctx.status = 200;
       ctx.body = user;
     } else {
-      ctx.status = 404;
+      throw new NotFoundException();
     }
   }
 
   public static async updateUser(ctx: any) {
+    const userId = +ctx.params.id;
+    if (userId !== +ctx.state.user.id) {
+      throw new ForbiddenException();
+      return;
+    }
+
     const userRepository = getManager().getRepository(User);
     await userRepository.update(+ctx.params.id, ctx.request.body);
     const updatedUser = await userRepository.findOne(+ctx.params.id);
@@ -37,6 +44,13 @@ export default class UserController {
   }
 
   public static async deleteUser(ctx: any) {
+    const userId = +ctx.params.id;
+
+    if (userId !== +ctx.state.user.id) {
+      throw new ForbiddenException();
+      return;
+    }
+
     const userRepository = getManager().getRepository(User);
     await userRepository.delete(+ctx.params.id);
